% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{build_diagnostics}
\alias{build_diagnostics}
\title{Build diagnostics list from a wicmad() result}
\usage{
build_diagnostics(res, Y = NULL, t = NULL)
}
\description{
Build diagnostics list from a wicmad() result
}
