% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_rows}
\alias{dist_rows}
\title{Compute Distance Matrix Between Rows}
\usage{
dist_rows(t)
}
\arguments{
\item{t}{Numeric vector or matrix of input coordinates}
}
\value{
Distance matrix between all pairs of rows
}
\description{
Computes pairwise distances between rows of input coordinates.
}
\examples{
dist_rows(c(1, 2, 3))  # 3x3 distance matrix
}
