% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eig_Bshape}
\alias{eig_Bshape}
\title{Eigendecomposition of Coregionalization Matrix}
\usage{
eig_Bshape(L, M)
}
\arguments{
\item{L}{Numeric lower triangular matrix}

\item{M}{Integer number of channels}
}
\value{
List with eigenvectors U and eigenvalues Lambda
}
\description{
Computes eigendecomposition of coregionalization matrix \eqn{B = L L^T = U \Lambda U^T}.
}
