% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eig_Kx}
\alias{eig_Kx}
\title{Eigendecomposition of Kernel Matrix}
\usage{
eig_Kx(Kx)
}
\arguments{
\item{Kx}{Numeric kernel matrix}
}
\value{
List with eigenvectors U and eigenvalues Lambda
}
\description{
Computes eigendecomposition of kernel matrix \eqn{K_x = U \Lambda U^T}.
}
