% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_dyadic_J}
\alias{ensure_dyadic_J}
\title{Ensure Dyadic Wavelet Decomposition Level}
\usage{
ensure_dyadic_J(P, J)
}
\arguments{
\item{P}{Integer signal length}

\item{J}{Integer or NULL number of decomposition levels}
}
\value{
Integer number of decomposition levels
}
\description{
Ensures the number of wavelet decomposition levels J is appropriate for signal length P.
}
\examples{
ensure_dyadic_J(64, NULL)  # Returns 6 (\eqn{\log_2(64)})
ensure_dyadic_J(64, 4)     # Returns 4
}
