% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invlogit}
\alias{invlogit}
\title{Inverse Logit Transformation}
\usage{
invlogit(z)
}
\arguments{
\item{z}{Numeric vector}
}
\value{
Numeric vector in (0,1)
}
\description{
Computes the inverse logit transformation: \eqn{\text{invlogit}(z) = 1/(1+e^{-z})}
}
\examples{
invlogit(0)  # Returns 0.5
invlogit(c(-2, 2))  # Returns c(0.12, 0.88)
}
