% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_bias}
\alias{k_bias}
\title{Bias Kernel}
\usage{
k_bias(t, s0)
}
\arguments{
\item{t}{Matrix of input locations}

\item{s0}{Bias scale parameter}
}
\value{
Bias kernel matrix
}
\description{
Implements a bias kernel for non-zero mean functions.
}
