% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_mat52}
\alias{k_mat52}
\title{Matérn 5/2 Kernel}
\usage{
k_mat52(t, l_scale)
}
\arguments{
\item{t}{Matrix of input locations}

\item{l_scale}{Length scale parameter}
}
\value{
Kernel matrix
}
\description{
Implements the Matérn 5/2 kernel function.
}
