% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_periodic}
\alias{k_periodic}
\title{Periodic Kernel}
\usage{
k_periodic(t, l_scale, period)
}
\arguments{
\item{t}{Matrix of input locations}

\item{l_scale}{Length scale parameter}

\item{period}{Period parameter}
}
\value{
Kernel matrix
}
\description{
Implements the periodic kernel function for modeling periodic patterns.
}
