% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_sqexp}
\alias{k_sqexp}
\title{Squared Exponential Kernel}
\usage{
k_sqexp(t, l_scale)
}
\arguments{
\item{t}{Matrix of input locations}

\item{l_scale}{Length scale parameter}
}
\value{
Kernel matrix
}
\description{
Implements the squared exponential (RBF) kernel function.
}
