% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_updates.R
\name{mh_update_kernel_eig}
\alias{mh_update_kernel_eig}
\title{Update Kernel Parameters via Metropolis-Hastings}
\usage{
mh_update_kernel_eig(k, params, kernels, t, Y_list, a_eta, b_eta)
}
\arguments{
\item{k}{Integer cluster index}

\item{params}{List of cluster parameters}

\item{kernels}{List of kernel configurations}

\item{t}{Numeric vector or matrix of input coordinates}

\item{Y_list}{List of observed curve matrices}

\item{a_eta}{Numeric shape parameter for eta prior}

\item{b_eta}{Numeric rate parameter for eta prior}
}
\value{
Updated params list
}
\description{
Performs Metropolis-Hastings updates for kernel hyperparameters using eigendecomposition.
}
