% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nloc}
\alias{nloc}
\title{Get Number of Locations}
\usage{
nloc(t)
}
\arguments{
\item{t}{Numeric vector or matrix of input coordinates}
}
\value{
Integer number of locations
}
\description{
Returns the number of input locations (rows for matrix, length for vector).
}
\examples{
nloc(c(1, 2, 3))  # Returns 3
nloc(matrix(1:6, ncol=2))  # Returns 3
}
