% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Plot diagnostics (returns a named list of ggplot objects)}
\usage{
plot_diagnostics(diag)
}
\description{
Plot diagnostics (returns a named list of ggplot objects)
}
