% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelets.R
\name{precompute_wavelets}
\alias{precompute_wavelets}
\title{Precompute Wavelet Transforms}
\usage{
precompute_wavelets(Y_list, wf, J, boundary)
}
\arguments{
\item{Y_list}{List of matrices, where each matrix represents a functional curve}

\item{wf}{Character string specifying the wavelet filter}

\item{J}{Integer specifying the number of decomposition levels}

\item{boundary}{Character string specifying boundary condition}
}
\value{
List of the same length as Y_list, where each element contains the wavelet transform results
}
\description{
Precomputes wavelet transforms for all curves in a list to avoid repeated computation during MCMC.
}
\details{
This function is called once at the beginning of the MCMC algorithm to precompute
all wavelet transforms.
}
\examples{
Y_list <- list(matrix(rnorm(128), nrow = 64, ncol = 2), 
               matrix(rnorm(128), nrow = 64, ncol = 2))
precomp <- precompute_wavelets(Y_list, wf = "la8", J = 4, boundary = "periodic")
}
