% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{project_curve}
\alias{project_curve}
\title{Project Curve onto Subspace}
\usage{
project_curve(Yi, mu, V)
}
\arguments{
\item{Yi}{Numeric matrix of curve data}

\item{mu}{Numeric vector of mean}

\item{V}{Numeric matrix of eigenvectors}
}
\value{
Numeric matrix of projected curve
}
\description{
Projects curve onto subspace defined by eigenvectors: \eqn{Y_{\text{proj}} = V V^T (Y - \mu) + \mu}
}
