% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{save_diagnostics_from_res}
\alias{save_diagnostics_from_res}
\title{One-liner: run model then save diagnostics to disk}
\usage{
save_diagnostics_from_res(
  res,
  Y = NULL,
  t = NULL,
  out_dir = "diagnostics",
  dpi = 150,
  width = 9,
  height = 6,
  units = "in"
)
}
\description{
One-liner: run model then save diagnostics to disk
}
