% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{save_plots_png}
\alias{save_plots_png}
\title{Save plots to PNG files}
\usage{
save_plots_png(
  plots,
  out_dir = "diagnostics",
  dpi = 150,
  width = 9,
  height = 6,
  units = "in"
)
}
\description{
Save plots to PNG files
}
