% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stick_to_pi}
\alias{stick_to_pi}
\title{Stick-Breaking to Probability Vector}
\usage{
stick_to_pi(v)
}
\arguments{
\item{v}{Numeric vector of stick-breaking weights in \link{0,1}}
}
\value{
Numeric vector of probabilities that sum to 1
}
\description{
Converts stick-breaking weights to probability vector using the stick-breaking process:
\eqn{\pi_k = v_k \prod_{j=1}^{k-1} (1-v_j)}
}
\examples{
stick_to_pi(c(0.5, 0.3, 0.8))  # Returns c(0.5, 0.15, 0.12)
}
