% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_v_given_z}
\alias{update_v_given_z}
\title{Update Stick Weights Given Cluster Assignments}
\usage{
update_v_given_z(v, z, alpha)
}
\arguments{
\item{v}{Numeric vector of stick weights}

\item{z}{Integer vector of cluster assignments}

\item{alpha}{Numeric concentration parameter}
}
\value{
Updated vector of stick weights
}
\description{
Updates stick-breaking weights given cluster assignments using Beta posteriors:
\eqn{v_k \sim \text{Beta}(1 + n_k, \alpha + \sum_{j=k+1}^K n_j)}
}
